trying to clean up file management. why is github so hard
