+++ agency_info.csv
@@ -1,0 +1,9 @@
+Agency ID, Short Name, Lat1, Lon1, Lat2, Lon2
+8,CHT,35.745483,-79.16646,35.97005,-79.00187
+12,GoTriangle,35.692123,-79.111105,36.087248,-78.323486
+16,NCSU Wolfline,35.768534,-78.716686,35.801764,-78.663816
+20,GoRaleigh,35.650967,-78.783672,35.990819,-78.50063
+24,GoDurham,35.884205,-78.99653,36.11652,-78.776801
+176,Duke Transit,35.994114,-78.955356,36.022726,-78.914718
+367,GoCary,35.74,-78.858,35.836,-78.698
+575,PART,35.7113703157,-80.6066894531,36.4082858756,-78.8757324219
