+++ route_info.csv
@@ -1,4 +1,4 @@
-Route ID,Short Name,Long Name
+Route ID,Short Name,Long Name,Stops
 4000122,WRX,Triangle Transit - Wake Forest-Raleigh Express (operated by GoRaleigh),4099990,4100206,4100218,4100314,4055098,4100426,4100434,4100198
 4000123,ZWX,Triangle Transit - Zebulon-Wendell-Raleigh Express (operated by GoRaleigh),4040178,4040190,4105850,4102458,4100206,4100218,4100426,4100434,4103262,4102326,4100118,4040182
 4000124,CRX,Triangle Transit - Chapel Hill-Raleigh Express,4062626,4062482,4061686,4100070,4197196,4100442,4100198,4062398,4100010,4105294,4199392,4143570,4062402,4101066
